requirements.txt
cryptography
argon2-cffi
rich
python-magic
pillow
opencv-python
soundfile
pydub
passlib
dnspython
idna
tldextract
scikit-learn
numpy

